Refactor common code
* Move "table scan" common code to `common-table-scan`
* Move event emitting logic and grouping (i.e. event size) to new
  `common` object
* Fix bugs with "new row" query
